requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
markdown>=3.5.0
tabulate>=0.9.0
lxml>=4.9.0
html2text>=2020.1.16
transformers>=4.35.0
torch>=2.0.0
sec-api>=1.0.0
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
python-magic>=0.4.27
pdfplumber>=0.9.0